requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
pycryptodome==3.10.1
    # via -r requirements.in
pygments==2.10.0
    # via rich
rich==10.7.0
    # via -r requirements.in
